Upgrade to Clap 4 and fix some error handling

+++ src/logging.rs
@@ -7,7 +7,7 @@ use pretty_env_logger::env_logger::fmt::Style;
 use pretty_env_logger::env_logger::fmt::StyledValue;
 use crate::APPLICATION_NAME;
 
-pub fn set_verbosity_value(value: usize, is_quiet: bool) {
+pub fn set_verbosity_value(value: u8, is_quiet: bool) {
     let env_var_name = format!("{}_LOG", APPLICATION_NAME.to_ascii_uppercase());
     let level = std::env::var(env_var_name).unwrap_or(String::from("INFO"));
     let mut level = level.as_str();
